restore
change the gre_q

+++ db/migrate/20160920211031_add_gre_q_to_students.rb
@@ -1,5 +1,5 @@
 class AddGreQToStudents < ActiveRecord::Migration
   def change
-    add_column :students, :gre_q, :int
+    add_column :students, :gre_q, :float
   end
 end
